pyproject.toml
[project]
name = "leanframe"
version = "0.2.0"
description = "LeanFrame is a pandas-like API built on top of Ibis."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas >=2.2.0",
    "pyarrow >=15.0.2",
    "ibis-framework>=11.0.0,<12.0.0",
    "pyopenssl>=26.1.0",
]

[dependency-groups]
dev = [
    "ibis-framework[bigquery,duckdb]>=10.6.0",
    "ipython>=8.37.0",
    "mypy>=1.17.1",
    "pandas-stubs>=2.3.0.250703",
    "pytest>=8.4.1",
    "ruff>=0.12.4",
    "twine>=6.2.0",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[tool.ruff]
exclude = ["scratch", "vault"]
